Add check for file extension of recent files
While we cannot prevent the user from opening malformed files, we can restrict the listing of those files that don't even end in the proper extension. In the rare event that users change the contents of this data file manually, files with unfitting extensions will simply not be shown.

The motivation for this is the principle of trusting user input as little as possible.