Remove trailing newlines of recent files
Unfortunately, the `File` constructor does not perform any sort of cleanup on the path it receives. Because of this, if I pass a complete line (including `\n`) to it as path, I get that same line with new-line character when I return `.getPath()`. This is not only bad for actually opening the file later, but it also causes issues in the UI, which is why each line needs to be trimmed before it is used as path.