Fix zooming at extremes
When the user scrolled further than the minimum or maximum, the internal radius property was continuously updated and kept decreasing/increasing, even though the actual visual property was not being updated. This meant that the user has to scroll quite a while to go back after an over-scrolling action. This PR fixes that, by moving the check of updating that property to another place.