Fix height calculation issue
Previously all nodes in the subgraph were used for determining the number of lanes. This caused weird behaviour for TB328 where surrounding, but non visible, nodes had a very high Y position, but all nodes in the current view did not. Thus, sometimes it looked like the graph was going off the screen, while in fact the number of lanes was so high that there was barely space for the top lane.