Rename 'Radius' to 'View Radius' in bookmark UI
One of the insights we gathered while interviewing test users was that the meaning of the radius property of a bookmark was not really clear to them. They did, however, find 'View Radius' a more clear description of what it meant, which is why I'm proposing to rename the relevant parts of the UI to reflect this.